Add quarterly reminders feature
This feature looked simple, but required changes that touched a few
different classes.  It was hard not to design something that could
simply send messages on specific dates (instead of being constrained
to the four quarter dates), but what is here can be easily adapted
to allow for scheduling messages on any date.

I tried to leave the existing models and service objects alone
as much as I could.  Now there are two types of messages: "onboarding"
which are those sent at intervals from an employee's start date,
and "quarterly" which are sent four times a year on specific dates.

Quarterly messages will repeat year after year, this was not
addressed in the requirements so I made an assumption.

Also a quarterly message will be sent to a new employee up to
one week after the quarterly message date; this was also an
assumption and easily undone or adjusted.

Some may object to the helper methods I scatter at the bottom
of my specs.  Some of them are repeated in different files.
This is no accident--I prefer to have, in a spec and in a test
example, all the contextual information at hand to describe
what is happening in the tests.  Then you can look at any
test and know from A to Z what is happening, without needing
to hunt for mystery values hidden in helpers, etc.

This is an alpha commit, I suppose.  feedback is welcome.